src/requirements.txt
alembic==1.14.1
annotated-types==0.7.0
asyncpg==0.30.0
greenlet==3.1.1
hiredis==3.1.0
Mako==1.3.9
MarkupSafe==3.0.2
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
redis==5.2.1
SQLAlchemy==2.0.38
typing_extensions==4.12.2
